Move some logic to new context file
